Refactor project into modular files
